requirements.txt
Django==5.0.2
django-storages==1.14.2
boto3==1.34.34
Pillow==10.2.0
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-cleanup==8.1.0
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-filter==23.5
django-widget-tweaks==1.5.0
